+++ src/Shared/google-analytics.tsx
@@ -5,7 +5,7 @@
  */
 
 import React, { useEffect } from "react"
-import ReactGA from "react-ga"
+import ReactGA from "react-ga4"
 import { useLocation } from "react-router-dom"
 
 interface Props {
@@ -19,8 +19,7 @@ const GAListener: React.FC<Props> = ({ children, trackingId }: Props) => {
   useEffect(() => {
     if (trackingId) {
       ReactGA.initialize(trackingId)
-      ReactGA.set({ page: location.pathname })
-      ReactGA.pageview(`${location.pathname}${location.search}`)
+      ReactGA.send({ hitType: "pageview", page: `${location.pathname}${location.search}` })
     }
   }, [location, trackingId])